Refactor code by extracting code related to the level of the wild Pokemon and moving it to the newly created WildPokemonLevel component. Add WildPokemonLevel component to CatchingSimulator component and page title (previously in CaptureChance component).
Add attribute hp to state of catchingSimulatorSlice, which is an object
with attributes currentHP and maximumHP. Add reducer setLevel, which
sets the sta.pokemon.level value along with the state.hp.maximumHP and
state.hp.currentHP values as the hp of the Pokemon changes depending on
the level.

Adjust attributes level, currentHP, and maximumHP in component
CaptureChance to read value from the catchingSimulator state.