Create new component SimulatorPokemonGender, which loads all genders the wild and user Pokemon can be as radio button options. Users are also able to change a Pokemon's gender if the Pokemon can be more than one gender.
Add reducer setGender, which receives the variables gender and isWild in
the payload. The isWild variable is a boolean that determines whether
the wildPokemon or userPokemon objects' gender attribute will be set to
the value of gender from the payload.

Import and implement SimulatorPokemonGender component in components
UserPokemonSettings and WildPokemonSettings.