Create new components DialogBox and PokemonEncounter.
Add initialState variable dialogBoxMessage and reducer
setDialogBoxMessage, which will set the dialogBoxMessage to the value
of 'message' passed in the action payload.

Component DialogBox displays a sentence one character at a time using
setInterval. Each time the message changes (read from the
catchingSimulator state variable 'message'), the component will render
again with the new message one character at a time as well.

Component PokemonEncounter displays the encounter UI, the Pokemon
encountered, and two buttons: Throw and Run/Return to Menu. Pressing the
throw button will call the function throwBall(). The action
setDialogBoxMessage will dispatch to reflect how many shakes have
occured and whether or not the catch was successful. Pressing the
Run/Return to menu button will end the encounter.

Add useState variable encounterInProgress in CatchingSimulator
component. This variable has an initial value of false and is used for
conditional rendering.

When false, components BattleSettings, WildPokemonSettings, and
UserPokemonSettings are displayed, allowing users to adjust how they want the encounter to be (changing of Pokemon
will be added to these group of settings when implemented).

When true, components related to the encounter such as PokemonEncounter,
PokeBallsList, SelectedBallCaptureChance, and AllBallsCaptureChances are
displayed.

Import and implement DialogBox in PokemonEncounter.

Import and implement PokemonEncounter in component CatchingSimulator.