Add captureRate attribute to action paylod of setCatchingPokemonID reducer, and set captureRate attribute of wildPokemon to value of captureRate.
Add conditional in the useEffect of the CatchingPokemonList component
so that the API to retrieve the list of all Pokemon is called only once when the variable pokemonList is null.

Move API calls getPokedexEntry and getSpeciesDetails and dispatch of
action setCatchingPokemonID from the useEffect to the onMouseDown event of the list items.